Fixing NCCH flag detection code for zerokey
Problem:
Encrypted CCIs no longer work with 3dsconv.py. All xorpads are incorrectly displayed as corrupt or a mismatch.

Analysis:
The recently-committed code that checked the NCCH flags for zerokey encryption was calling read(1) twice for each flag comparison, which resulted in the wrong byte being used for comparison in the second case. As a result, all encrypted CCIs are incorrectly flagged as being encrypted with a fixed/zero key.

This commit stores the bitmask in a temporary variable instead. Encrypted CCIs are now properly detected and packed as expected.